Move all the event handling to EvenHandling
